pyproject.toml
[project]
name = "bml-django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-extensions>=4.1",
    "factory-boy>=3.3.3",
    "ipython>=9.12.0",
    "psycopg[binary]>=3.3.3",
    "ptpython>=3.0.32",
]
